+++ src/main/java/com/coniverse/dangjang/domain/guide/bloodsugar/guideformat/format/CautionGuideFormat.java
@@ -1,4 +1,4 @@
-package com.coniverse.dangjang.domain.guide.bloodsugar.guideformat;
+package com.coniverse.dangjang.domain.guide.bloodsugar.guideformat.format;
 
 import java.util.Arrays;
