+++ src/main/java/com/coniverse/dangjang/domain/guide/bloodsugar/guideformat/format/GuideFormat.java
@@ -1,4 +1,4 @@
-package com.coniverse.dangjang.domain.guide.bloodsugar.guideformat;
+package com.coniverse.dangjang.domain.guide.bloodsugar.guideformat.format;
 
 public interface GuideFormat {
 	String getTitle();
